+++ src/app/group/[id]/page.tsx
@@ -7,21 +7,19 @@ import { board } from '@/types/boards';
 import ContentPhoto from '@/component/ContentPhoto';
 import { useRecoilState } from 'recoil';
 import { themeState } from '@/recoil/theme';
-import Drawer from '@/component/Drawer';
-import StrcatHeader from '@/component/StrcatHeader';
 import BottomButton from '@/component/BottomButton';
 import { observeState } from '@/recoil/observe';
 import StrcatGroupTitle from '@/component/StrcatGroupTitle';
 import { scrollToAdd, setMap } from '@/utils/scrollTo';
 import { useRouter } from 'next/navigation';
 import ShortCut from '@/component/Icon/ShortCut';
-import { content } from '@/types/content';
 import ShareButton from '@/component/ShareButton';
 import { useLogin } from '@/hooks/useLogin';
 import Loading from '@/component/Loading';
 import CatAnimation from '@/component/CatAnimation';
 import { useCat } from '@/hooks/useCat';
 import { catAction } from '@/types/animation';
+import HeaderLayout from '@/component/HeaderLayout';
 
 export default function Group({ params }: { params: { id: string } }) {
   const [title, setTitle] = useState<string>('');
@@ -89,8 +87,7 @@ export default function Group({ params }: { params: { id: string } }) {
 
   return (
     <div className={`${theme.background}  min-h-full`}>
-      <Drawer />
-      <StrcatHeader />
+      <HeaderLayout />
       <CatAnimation />
       <div
         className={`relative w-full py-[24px] pt-[56px] ${theme.background}
