+++ src/component/HeaderLayout/DrawerItem.tsx
@@ -1,5 +1,3 @@
-import Image from 'next/image';
-
 interface Props {
   title: string;
   icon: React.ReactElement;
