+++ src/component/HeaderLayout/index.tsx
@@ -1,0 +1,11 @@
+import Drawer from './Drawer';
+import DefaultHeader from './DefaultHeader';
+
+export default function HeaderLayout() {
+  return (
+    <>
+      <DefaultHeader />
+      <Drawer />
+    </>
+  );
+}
