+++ src/types/drawerBoard.ts
@@ -1,0 +1,4 @@
+export interface drawerBoard {
+  id: string;
+  title: string;
+}
